New build script using ostree checkout
This script uses a checkout of the currently released stable build to
generate the HTML documentation exactly matching what's in the OS. The
build is then done in a checkout using bwrap so it can all be done
unprivileged. Additionally, a manifest of the current commit and ref are
kept to track when a build is actually needed.