generate-html-docs: Fix check for yelp-build's presence
It looks like the '--help' argument has been dropped from yelp-build, as
'yelp-build --help' now prints an error and exit with status of 1,
breaking our check.

The shell building 'command' is documented to exit with a status of 0 if
the command was found and 1 otherwise, when used with -v, so this seems
a safe bet.

https://phabricator.endlessm.com/T33073