Build and publish per-branch subdirs
Previously the help center for a particular EOS branch would be
published to the root of the S3 bucket so that it formed the entire
website. Instead, we want to publish different EOS branches to
subdirectories so that multiple branches can be hosted on one website.
The root index pointing to the available branches will be handled in the
deployment.

Add `--branch` options to the scripts so they operate on branched
subdirectories. This is particularly important for the publishing script
since it needs to only synchronize the subdirectory and not the entire
bucket.

https://phabricator.endlessm.com/T32822